Blindly assume the user's terminal can do Unicode or be switched into it.
I expect this to work in 99% cases, and if it fails, there's probably no
regression since I don't know of a terminal that can be easily switched to
CP437 but doesn't support "\e%G".

It would be better to check if the terminal uses Unicode and convert
everything to ASCII, and I'll do that soon, but this version might be
good enough for now.